fix: eslint + my minor changes
