+++ server/src/helpers/helpers.js
@@ -1,0 +1,9 @@
+export const filterFields = (body, allowedFields) => {
+    const filtered = {};
+    Object.keys(body).forEach(key => {
+        if (allowedFields.includes(key)) {
+            filtered[key] = body[key];
+        }
+    });
+    return filtered;
+};
