+++ server/src/index.js
@@ -1,7 +1,7 @@
 import express from 'express';
 import cors from 'cors';
 import userRoutes from './routes/userRoutes.js';
-import authRoutes from "./routes/authRoutes.js";
+import authRoutes from './routes/authRoutes.js';
 import serviceRoutes from './routes/serviceRoutes.js';
 import orderRoutes from './routes/orderRoutes.js';
 import driver from './config/db.js';
@@ -23,40 +23,40 @@ app.use('/api/orders', orderRoutes);
 const PORT = 3000;
 
 const initDatabase = async () => {
-  const session = driver.session();
-  try {
-    const constraints = [
-      'CREATE CONSTRAINT unique_user_id IF NOT EXISTS FOR (u:User) REQUIRE u.user_id IS UNIQUE',
-      'CREATE CONSTRAINT unique_order_id IF NOT EXISTS FOR (o:Order) REQUIRE o.order_id IS UNIQUE',
-      'CREATE CONSTRAINT unique_service_id IF NOT EXISTS FOR (s:Service) REQUIRE s.service_id IS UNIQUE',
-      'CREATE CONSTRAINT unique_history_id IF NOT EXISTS FOR (h:StatusHistory) REQUIRE h.history_id IS UNIQUE',
-      'CREATE CONSTRAINT unique_log_id IF NOT EXISTS FOR (l:ImportExportLogs) REQUIRE l.log_id IS UNIQUE',
-      'CREATE CONSTRAINT unique_user_email IF NOT EXISTS FOR (u:User) REQUIRE u.email IS UNIQUE'
-    ];
+    const session = driver.session();
+    try {
+        const constraints = [
+            'CREATE CONSTRAINT unique_user_id IF NOT EXISTS FOR (u:User) REQUIRE u.user_id IS UNIQUE',
+            'CREATE CONSTRAINT unique_order_id IF NOT EXISTS FOR (o:Order) REQUIRE o.order_id IS UNIQUE',
+            'CREATE CONSTRAINT unique_service_id IF NOT EXISTS FOR (s:Service) REQUIRE s.service_id IS UNIQUE',
+            'CREATE CONSTRAINT unique_history_id IF NOT EXISTS FOR (h:StatusHistory) REQUIRE h.history_id IS UNIQUE',
+            'CREATE CONSTRAINT unique_log_id IF NOT EXISTS FOR (l:ImportExportLogs) REQUIRE l.log_id IS UNIQUE',
+            'CREATE CONSTRAINT unique_user_email IF NOT EXISTS FOR (u:User) REQUIRE u.email IS UNIQUE'
+        ];
     
-    if (process.env.NODE_ENV === 'development') {
-        await seedDatabase();
-    }
+        if (process.env.NODE_ENV === 'development') {
+            await seedDatabase();
+        }
 
-    for (const query of constraints) {
-      await session.run(query);
-      console.log(`Constraint created: ${query.substring(0, 60)}...`);
+        for (const query of constraints) {
+            await session.run(query);
+            console.log(`Constraint created: ${query.substring(0, 60)}...`);
+        }
+        console.log('Database constraints initialized');
+    } catch (error) {
+        console.error('Database initialization error:', error);
+    } finally {
+        await session.close();
     }
-    console.log('Database constraints initialized');
-  } catch (error) {
-    console.error('Database initialization error:', error);
-  } finally {
-    await session.close();
-  }
 };
 
 app.listen(PORT, async () => {
-  console.log(`Server running on http://localhost:${PORT}`);
-  await initDatabase();
+    console.log(`Server running on http://localhost:${PORT}`);
+    await initDatabase();
 });
 
 process.on('SIGINT', async () => {
-  await driver.close();
-  console.log('Neo4j connection closed');
-  process.exit(0);
+    await driver.close();
+    console.log('Neo4j connection closed');
+    process.exit(0);
 });
